+++ day_iii/power_meter.c
@@ -43,23 +43,23 @@ int main(int argc, char **argv)
     uint8_t **readings = get_input(argv[1], &reason, &field_width, &count);
 
     // Try to get the power consumption
-    int power = get_power_consumption(readings, &count, &field_width);
+    size_t power = get_power_consumption(readings, &count, &field_width);
 
     if (0 == power) {
         fprintf(stderr, "[ERROR] Failed to determine power consumption!\n");
     } else {
         // Report the power consumption
-        fprintf(stdout, "[*] Power consumption: %d\n", power);
+        fprintf(stdout, "[*] Power consumption: %lu\n", power);
     }
 
     // Fetch the life support rating
-    int life_support = get_life_support(readings, &count, &field_width);
+    size_t life_support = get_life_support(readings, &count, &field_width);
 
     if (0 == life_support) {
         fprintf(stderr, "[ERROR] Failed to determine life support rating!\n");
     } else {
         // Report the life support rating
-        fprintf(stdout, "[*] Life support rating: %d\n", life_support);
+        fprintf(stdout, "[*] Life support rating: %lu\n", life_support);
     }
 
     // Cleanup
